add ckb-auth-cli
commit 4f0225032cd5248597168ac786bd52bc625b306c
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 21:45:11 2023 +0800

    generate random litecoin key

commit 9f69723bfa9c292b1004bd213c824ddc4dd36c65
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 21:41:12 2023 +0800

    Revert "verify only the simplest transaction for litecoin"

    This reverts commit 004bd45c1db268ed2b5d9f2f38184c32200b2df6.

commit 98c83df6701f12721597a63faae90996b096d590
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 21:41:04 2023 +0800

    Revert "use deterministic rng for debugging"

    This reverts commit d4ab023a266b8d4589087dc30652cc5583fd6981.

commit 4fbc5e46d2a7c0ae93b5f188694a075b8be07293
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 21:36:32 2023 +0800

    remove debug info

commit 41da293d6cd00e967576fde2da3c05dcbb7fbdde
Merge: 93b2ab6 94a52a1
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 21:34:07 2023 +0800

    Merge branch 'ckb-auth-cli-debug-litecoin-pubkey-hash' into ckb-auth-cli

commit 94a52a166aadd9e4d256df8c02de2abb6c652a54
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 21:28:49 2023 +0800

    don't double hash while generate litecoin message to sign

commit 300864de0f3dcf209a8814036be0d42d5e0f3f1b
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 18:58:25 2023 +0800

    parse blockchain in ckb-auth-cli

commit d4ab023a266b8d4589087dc30652cc5583fd6981
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 18:57:51 2023 +0800

    use deterministic rng for debugging

commit 004bd45c1db268ed2b5d9f2f38184c32200b2df6
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 17:56:56 2023 +0800

    verify only the simplest transaction for litecoin

commit 71ff9fd455d08fffaafa1772c530c65bdc96d53e
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 17:52:21 2023 +0800

    adjust to new path

commit 581fd957c8a5b54371732cc947e82b1e6a02bdc1
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 15:01:24 2023 +0800

    remove old test files

commit 82494f1ba86c21ef13bdee4539b4077f530bc056
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 15:34:29 2023 +0800

    debug pub keys in ckb-auth

commit 93b2ab658f214f797442e348572501eb0cc40e4b
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 15:01:24 2023 +0800

    verify only the simplest transaction for litecoin

commit 703d44a04a28f73c58c1ee725c4e3e9acc5ee271
Author: YI <uuuuuu@protonmail.com>
Date:   Fri Jun 9 14:28:42 2023 +0800

    generate litecoin key deteministically

commit 170ec002603d4bdb8cca55cf04dd75497523890b
Author: YI <uuuuuu@protonmail.com>
Date:   Thu Jun 8 18:50:36 2023 +0800

    pass the right signature to verify

commit fab51470773bb9ac4ac99eb1f3e46bb59b792eeb
Author: YI <uuuuuu@protonmail.com>
Date:   Thu Jun 8 16:38:19 2023 +0800

    encode signatures with hex or base64

commit 079f960785289c4c19af654f8b397c3cc799d159
Author: YI <uuuuuu@protonmail.com>
Date:   Thu Jun 8 16:17:20 2023 +0800

    generate transaction deterministically

commit 29d751fb3db7e96675b59f80934d351f338b835f
Author: YI <uuuuuu@protonmail.com>
Date:   Thu Jun 8 14:42:27 2023 +0800

    load pub key hash

commit 353dc58887518c2fb1cfb24ac8c0d4d485eef151
Author: YI <uuuuuu@protonmail.com>
Date:   Thu Jun 8 13:13:10 2023 +0800

    get public key hash

commit 85759e551d8b363b080d491b896d44684a0d6763
Author: YI <uuuuuu@protonmail.com>
Date:   Thu Jun 8 12:16:33 2023 +0800

    remove a few unused parameters

commit 5fa648047e823e397db2da19b1ed2532d726ad91
Author: YI <uuuuuu@protonmail.com>
Date:   Thu Jun 8 12:10:19 2023 +0800

    fix warnings

commit 0a020fe19f78f421bb4c779a2d818956949ce910
Author: YI <uuuuuu@protonmail.com>
Date:   Thu Jun 8 12:07:03 2023 +0800

    add pubkey for BitcoinAuth for get_pub_key_hash only like operations

commit e3ea56f77a4dfa2a4701b108937c2c4660db7842
Author: YI <uuuuuu@protonmail.com>
Date:   Tue Jun 6 17:46:03 2023 +0800

    get public key hash

commit 10f6fe4c5641e2bacdc1146235458ea4907153a4
Author: YI <uuuuuu@protonmail.com>
Date:   Tue Jun 6 17:44:39 2023 +0800

    set signature and verify

commit 1783d1595c010a42e85bbf53acf7e311a370aca2
Author: YI <uuuuuu@protonmail.com>
Date:   Tue Jun 6 17:20:19 2023 +0800

    get message to sign

commit 9e6a1e2af662bc34d97d0232f5bd5d0cec4bf925
Author: YI <uuuuuu@protonmail.com>
Date:   Tue Jun 6 16:52:58 2023 +0800

    generate message for bitcoin

commit e321bf25141da4303d882c7d3a76f3185aabb124
Author: YI <uuuuuu@protonmail.com>
Date:   Tue Jun 6 16:39:39 2023 +0800

    move testing code to lib.rs so that ckb-auth-cli can use it

commit bf9ebaf43af5023a1a2a93fc633abc4eb40ac797
Author: YI <uuuuuu@protonmail.com>
Date:   Tue Jun 6 16:33:15 2023 +0800

    add skeleton code for ckb-auth-cli command

commit bc1e0a33741aea50ff1ee88665973a900062e3ba
Author: YI <uuuuuu@protonmail.com>
Date:   Tue Jun 6 16:01:31 2023 +0800

    add ckb-auth-cli binary
